+++ inc/sigfox_ep_lib_version.h
@@ -43,7 +43,7 @@
 
 /*** Main version ***/
 
-#define SIGFOX_EP_LIB_VERSION			"v3.3"
+#define SIGFOX_EP_LIB_VERSION			"v3.4"
 
 /*** Compilation flags ***/