+++ src/manuf/mcu_api.c
@@ -190,7 +190,7 @@ MCU_API_status_t MCU_API_get_voltage_temperature(sfx_u16 *voltage_idle_mv, sfx_u
 }
 #endif
 
-#ifdef CERTIFICATION
+#if (defined CERTIFICATION) && (defined BIDIRECTIONAL)
 /*******************************************************************/
 MCU_API_status_t MCU_API_print_dl_payload(sfx_u8 *dl_payload, sfx_u8 dl_payload_size, sfx_s16 rssi_dbm) {
 	/* To be implemented by the device manufacturer */